Move YAML loading under rollouts
